feat: add sorting to games page
The fix to the `getIconByCaseInsensitiveName()` function was necessary as using `String.replace("-", "")` only removes the first instance of the hyphen, but `String.replaceAll("-", "")` removes all instances of the hyphen.

The rest of the patch adds a sorting dropdown to the games page, whilst preserving the search functionality.

This is one pull request in a stack... they are all based on top of each other, so the order is important.

The advantage of this approach is that you are guaranteed no conflicts and reviews should be shorter.

Review each patch and merge before you move on to the next, this way, nothing will break and the code will be easier to review.

Start from the bottom (above main) and work your way up the stack until you reach the top.

* **#33** 👈
* **#32**
* **#31**
* **#30**
* **#29**
* **#28**
* **#22**
* **#27**
* **#26**
* **#25**
* **main**